Add files via upload
First enemy type is the minion and uses 3 different types of weapons - note that it uses the collision of the weapon so with the animations it can only hit the player if the player overlaps with the weapon itself. Ranger enemy exists but behavior tree needs to be setup and whether it can shoot or not needs to be set up also.